requirements.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
dialogflow==0.5.1
Flask==1.0.2
future==0.16.0
google-api-core==1.4.0
google-auth==1.5.1
googleapis-common-protos==1.5.3
grpcio==1.15.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.2
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pymessenger==0.0.7.0
pytz==2018.5
requests==2.19.1
requests-toolbelt==0.8.0
rsa==4.0
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
